requirements.txt
appdirs==1.4.3
APScheduler==3.5.3
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bottle==0.12.13
certifi==2017.4.17
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
configparser==3.5.0
contextlib2==0.5.5
coverage==4.5.1
cryptography==2.3.1
decorator==4.3.0
Django==1.11.2
django-appconf==1.0.2
django-apscheduler==0.2.13
django-autoslug==1.9.3
django-compressor==2.2
django-crontab==0.7.1
django-debug-toolbar==1.10.1
django-haystack==2.8.1
django-ipware==2.1.0
django-local-settings==1.0b9
django-pagedown==1.0.6
django-uuslug==1.1.8
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
DjangoRestless==0.0.10
enum34==1.1.6
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
idna==2.5
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
itchat==1.2.32
jieba==0.39
jsonpickle==1.0
lazy-object-proxy==1.3.1
markdown2==2.3.6
mccabe==0.6.1
mistune==0.8.4
MySQL-python==1.2.5
olefile==0.46
packaging==18.0
pathlib2==2.3.2
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==1.0.15
pycparser==2.19
pyExcelerator==0.6.4.1
Pygments==2.2.0
PyJWT==1.5.0
pylint==1.9.3
PyMySQL==0.9.2
pyparsing==2.3.0
pypng==0.0.18
PyQRCode==1.2.1
python-memcached==1.59
python-slugify==1.2.6
pytz==2017.2
raven==6.9.0
rcssmin==1.0.6
requests==2.17.3
rjsmin==1.0.12
scandir==1.9.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
tzlocal==1.5.1
Unidecode==1.0.22
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
WeRoBot==1.6.0
Whoosh==2.7.4
win-unicode-console==0.5
wrapt==1.10.11
wxpy==0.3.9.8
xlrd==1.1.0
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.11.0
